+++ lemon/extradmin/__init__.py
@@ -1,6 +1,9 @@
-from lemon.extradmin.options import ModelAdmin
+from lemon.extradmin.options import ModelAdmin, AppAdmin
 from lemon.extradmin.options import StackedInline, TabularInline
 from lemon.extradmin.sites import AdminSite, site
 
 
-__all__ = ['ModelAdmin', 'StackedInline', 'TabularInline', 'AdminSite', 'site']
+__all__ = [
+    'ModelAdmin', 'AppAdmin', 'StackedInline', 'TabularInline', 'AdminSite',
+    'site',
+]